Redesign Multiplexer - reduce number of used gates from 14 to 4.
Instead of falling asleep I kept thinking about Mux chip.
Inspired of how AND and OR work in JavaScript variable assigments I came up with very simple formula:
mux(a, b, sel) = (a && !sel) || (b && sel)
In JS it could be for example:
var mux = (!sel && a) || (sel && b)
Then I have realised that this is actually simplification of my current design, which was just canonical form of mux boolean function.
In a result of distributivity of AND over OR one can derive above simple formula from canonical form:
(a && b && !sel) || (a && !b && !sel) || (b && a && sel) || (b && !a && sel)
(a && !sel) && (b || !b) || (b && sel) && (a || !a)
(a && !sel) || (b && sel)
Q.E.D